is  prime prog final

+++ is_prime_prog/sources.h
@@ -5,7 +5,7 @@
 #include <iostream>
 
 
-bool is_prime(long num);
+bool is_prime(long long int num);
 void printNum(long num);
 
 #endif /* SOURCES */
